Crate iced_glyphon

Glyphon provides a simple way to render 2D text with wgpu, cosmic-text and etagere.

Re-exports

pub use cosmic_text;

pub use cosmic_text::fontdb;

pub use cosmic_text::Action;

pub use cosmic_text::Affinity;

pub use cosmic_text::Attrs;

pub use cosmic_text::AttrsList;

pub use cosmic_text::AttrsOwned;

pub use cosmic_text::Buffer;

pub use cosmic_text::BufferLine;

pub use cosmic_text::CacheKey;

pub use cosmic_text::Color;

pub use cosmic_text::Command;

pub use cosmic_text::Cursor;

pub use cosmic_text::Edit;

pub use cosmic_text::Editor;

pub use cosmic_text::Family;

pub use cosmic_text::FamilyOwned;

pub use cosmic_text::Font;

pub use cosmic_text::FontSystem;

pub use cosmic_text::LayoutCursor;

pub use cosmic_text::LayoutGlyph;

pub use cosmic_text::LayoutLine;

pub use cosmic_text::LayoutRun;

pub use cosmic_text::LayoutRunIter;

pub use cosmic_text::Metrics;

pub use cosmic_text::ShapeGlyph;

pub use cosmic_text::ShapeLine;

pub use cosmic_text::ShapeSpan;

pub use cosmic_text::ShapeWord;

pub use cosmic_text::Shaping;

pub use cosmic_text::Stretch;

pub use cosmic_text::Style;

pub use cosmic_text::SubpixelBin;

pub use cosmic_text::SwashCache;

pub use cosmic_text::SwashContent;

pub use cosmic_text::SwashImage;

pub use cosmic_text::Weight;

pub use cosmic_text::Wrap;

Structs

Cache

Resolution

The screen resolution to use when rendering text.

TextArea

A text area containing text to be rendered along with its overflow behavior.

TextAtlas

An atlas containing a cache of rasterized glyphs that can be rendered.

TextBounds

Controls the visible area of the text. Any text outside of the visible area will be clipped.

TextRenderer

A text renderer that uses cached glyphs to render text into an existing render pass.

Viewport

Enums

ColorMode

The color mode of an [Atlas].

PrepareError

An error that occurred while preparing text for rendering.

RenderError

An error that occurred while rendering text.